requirements.txt
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
twilio==6.29.3
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
